refactor(server): extract Tuist.Accounts.AgentAuth
Move the auth.md agent-registration workflow (JWT/JWKS verification, JTI
replay protection, claim ceremony, OTP generation, credential issuance,
revocation, audit-event recording) out of Tuist.Accounts into a focused
Tuist.Accounts.AgentAuth module. Tuist.Accounts keeps the existing
public API via thin delegators so existing callers and tests are
unchanged. Cuts ~900 LOC of agent-auth code from the broader Accounts
context and makes future changes to agent auth easier to scope and review.

Also fixes resend_agent_registration_claim so a nil email no longer raises
in validate_agent_registration_email — claim requests without an email
fall through to the stored registration email, while malformed emails
still return {:error, :invalid_email} (paired with the controller's new
invalid_email branch).
