+++ server/test/test_helper.exs
@@ -45,6 +45,7 @@ Mimic.copy(System)
 Mimic.copy(Time)
 Mimic.copy(Tuist.API.Pipeline)
 Mimic.copy(Tuist.Accounts)
+Mimic.copy(Tuist.Accounts.AgentAuth)
 Mimic.copy(Tuist.AppBuilds)
 Mimic.copy(Tuist.Authentication)
 Mimic.copy(Tuist.Authorization)
